+++ lib/Data/Time/Format/Parse.hs
@@ -12,6 +12,7 @@ module Data.Time.Format.Parse
     , module Data.Time.Format.Locale
     ) where
 
+import Control.Applicative ((<|>))
 import Control.Monad.Fail
 import Data.Char
 import Data.Proxy
@@ -148,7 +149,10 @@ instance Read ZonedTime where
     readsPrec n = readParen False $ \s -> [(ZonedTime t z, r2) | (t, r1) <- readsPrec n s, (z, r2) <- readsPrec n r1]
 
 instance Read UTCTime where
-    readsPrec n s = [(zonedTimeToUTC t, r) | (t, r) <- readsPrec n s]
+    readsPrec n s = do
+        (lt, s') <- readsPrec n s
+        (tz, s'') <- readsPrec n s' <|> pure (utc, s')
+        return (localTimeToUTC tz lt, s'')
 
 instance Read UniversalTime where
     readsPrec n s = [(localTimeToUT1 0 t, r) | (t, r) <- readsPrec n s]