+++ changelog.md
@@ -3,6 +3,7 @@
 ## [1.11]
 - add instance Ord DayOfWeek
 - add instance Read DiffTime (and NominalDiffTime)
+- change instance Read UTCTime to allow omitted timezone
 
 ## [1.10]
 - remove deprecated functions parseTime, readTime, readsTime